fix(labs): keep student packages dependency-free
