+++ .trellis/spec/guides/cross-platform-thinking-guide.md
@@ -1,0 +1,15 @@
+# 跨平台与独立分发检查清单
+
+当代码会在 Windows、Linux、macOS、CI 临时目录或脱离仓库的分发包中运行时，提交前检查：
+
+- “独立运行”必须把产物复制到仓库根目录之外再执行；Node 会沿父目录查找 `node_modules`，仓库内测试可能被根依赖假性兜底。
+- 分发包的运行时入口只静态加载包内文件和 Node 内置模块。脚手架、仓库扫描、发布等作者命令若依赖第三方包，应按命令动态加载，不能污染学习者运行路径。
+- Windows、GCC、Clang 的成功只能证明各自路径；Make、shell、路径空格、大小写与行尾行为要由对应平台测试覆盖，不能互相替代。
+- 临时目录测试同时覆盖普通路径与含空格路径，并在 `finally` 中精确清理自己创建的目录。
+- 修改用户可见文案、路径或编号时，搜索构建检查、发现测试和浏览器测试中的旧断言；测试应验证当前合同，而不是把历史展示形式固化成需求。
+
+相关规范：
+
+- [Lab 工具与学生包](../content/lab-tooling.md)
+- [跨层变更检查清单](cross-layer-thinking-guide.md)
+- [验证与 GitHub Pages](../quality/validation-and-pages.md)
