+++ .trellis/spec/guides/index.md
@@ -4,5 +4,6 @@
 
 - [code-reuse-thinking-guide.md](code-reuse-thinking-guide.md)：先找单一数据源和 VitePress 原生能力。
 - [cross-layer-thinking-guide.md](cross-layer-thinking-guide.md)：追踪 Markdown 从源文件到 Pages 点击的完整链路。
+- [cross-platform-thinking-guide.md](cross-platform-thinking-guide.md)：验证跨系统入口与脱离仓库的独立分发包，防止根依赖掩盖缺陷。
 
 具体实现规则从 [项目规范总索引](../index.md) 进入。
