+++ .trellis/tasks/archive/2026-08/08-31-lab-stable-id-numbering/design.md
@@ -87,6 +87,8 @@ README.labId
 
 树、面板、历史和统计均使用 `id` 查询进度，但导航仍可同时按 `id` 或目录名解析。
 
+学生分发包的运行时只加载无第三方依赖的稳定 ID 解析模块。仓库级 `new`、`locate` 按需加载扫描与脚手架依赖，避免独立学生包被根仓库的 `gray-matter` 或 `node_modules` 隐式绑住。
+
 ## 8. 并发、失败与回滚
 
 本地扫描无法感知另一未合并分支。重复 ID 由全仓内容校验和 CI 阻断，后合并分支同步 main 后重新创建或使用迁移工具修正未发布 Lab。
